refactor(view): プラン候補取得に関するhooksを取得
